+++ CHANGELOG.md
@@ -1,5 +1,19 @@
+## 2.2.0
++ Implement `endConnection` method to declaratively finish observer in iOS.
++ Remove `addTransactionObserver` in IAPPromotionObserver.m for dup observer problems.
++ Automatically startPromotionObserver in `initConnection` for iOS.
+
+## 2.1.5
++ Fix ios failed purchase handling problem in 11.4+ [#176](https://github.com/dooboolab/flutter_inapp_purchase/pull/176)
+
+## 2.1.4
++ Fix dart side expression warning [#169](https://github.com/dooboolab/flutter_inapp_purchase/pull/169).
+
+## 2.1.3
++ Fix wrong introductory price number of periods [#164](https://github.com/dooboolab/flutter_inapp_purchase/pull/164).
+
 ## 2.1.2
-+ Add PurchasedItem.transactionStateIOS.
++ Trigger purchaseUpdated callback when iap purchased [#165](https://github.com/dooboolab/flutter_inapp_purchase/pull/165).
 
 ## 2.1.1
 + Renamed `finishTransactionIOS` argument `purchaseToken` to `transactionId`.